requirement.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cssselect2==0.5.0
Django==4.0.3
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-easy-pdf==0.1.1
django-js-asset==2.0.0
django-nepali-datetime-field==0.5
django-xhtml2pdf==0.0.4
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
nepali==0.5.0
nepali-datetime==1.0.7
Pillow==9.0.1
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.1
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
windows-curses==2.3.0
xhtml2pdf==0.2.6
xlwt==1.3.0
